[Refactor] Player Balance Attribute Transfer
플레이어에서의 매직 넘버로 설정되어 있었던 수치를 전부 SerializeField로 수정하여 인스펙터 상으로 수치 조절 가능하게 함.

+++ Assets/Scripts/Managers/GameManager.cs
@@ -170,7 +170,7 @@ public class GameManager : Singleton<GameManager>
     /// 플레이어 쿨타임 설정(시간)
     /// </summary>
     /// <param name="time"></param>
-    public void SetCoolTime(int time)
+    public void SetCoolTime(float time)
     {
         m_playerSpellCoolTime = time;
     }
